Move job DTOs to correct package.
