+++ mico-core/src/main/java/io/github/ust/mico/core/dto/response/MicoApplicationJobStatusResponseDTO.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.github.ust.mico.core.dto;
+package io.github.ust.mico.core.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
